requirements.txt
tensorflow>=2.15
tensorflow-datasets>=4.9
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
umap-learn>=0.5
